requirements.txt
# Core dependencies for AI-Prepping
mlx>=0.11.0
mlx-lm>=0.10.0
faiss-cpu>=1.8.0
numpy>=1.24.0,<2.0.0
urllib3>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
pyyaml>=6.0
tiktoken>=0.5.0
rich>=13.0.0
click>=8.1.0
psutil>=5.9.0
sentence-transformers>=2.2.0

# Web UI (optional)
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0